...G....S. [DEV-2260] remove redundant casts from C code
# Conflicts:
#	src/libs/zbxdb/db.c
#	src/libs/zbxmemory/memalloc.c
#	src/libs/zbxnum/num.c
#	src/zabbix_agent/listener.c

+++ src/libs/zbxdbhigh/trigger_linking.c
@@ -1281,7 +1281,7 @@ static int	execute_triggers_inserts(zbx_vector_trigger_copies_insert_t *trigger_
 		zbx_db_insert_add_values(&db_insert, triggerid, trigger_copy_template->description,
 				(int)trigger_copy_template->priority, (int)trigger_copy_template->status,
 				trigger_copy_template->comments, trigger_copy_template->url,
-				(int)trigger_copy_template->type, (int)TRIGGER_VALUE_OK, (int)TRIGGER_STATE_NORMAL,
+				(int)trigger_copy_template->type, TRIGGER_VALUE_OK, TRIGGER_STATE_NORMAL,
 				trigger_copy_template->templateid, (int)trigger_copy_template->flags,
 				(int)trigger_copy_template->recovery_mode, (int)trigger_copy_template->correlation_mode,
 				trigger_copy_template->correlation_tag, (int)trigger_copy_template->manual_close,
